starlark: establish starlark-dialect-build-targets crate
The plan for implementing support for building installers is to
have this code be somewhat generic and not specific to PyOxidizer.
In order to avoid the dependency on PyOxidizer, we need common
Starlark code to live in its own crate. That common code should
be limited to the build system like primitives for defining targets.

This commit begins the process of extracting that Starlark code to
its own crate.

We start by moving the types that define targets. There are some
PyOxidizer implementation details in here. But hopefully we'll
tease those out in time.