starlark: make BuildContext a trait
The generic crate should make minimal assumptions about how a target's
build context is defined. This commit converts BuildContext to a trait
which enables implementations to provide their own context.

I'm not yet sure if this is flexible enough. But at least we get the
PyOxidizer specific pieces out of the generic crate. I suspect we may
need to undo this and move storage back to this type but implement
storage generically. We'll see.