starlark: use build targets crate for targets context
This commit is a big step towards using the new build targets crate.

We remove the storage of EnvironmentContext from PyOxidizer's
and remove the target symbols from PyOxidizer.

We now have 2 contexts stored in the Starlark environment: 1 for the
build targets and 1 for PyOxidizer.

There's still some cleanup to do. But this gets us mostly ported
over to the new crate.