tugger: add project scaffolding
Last year I created a proof-of-concept tool/crate called Tugger.
Its canonical home is https://github.com/indygreg/tugger. The purpose
of Tugger was to hold generic application packaging/distribution
functionality.

At the time, Tugger was an experiment to play around with Starlark,
before PyOxidizer was using Starlark.

I haven't touched Tugger in a very long time.

Anyway, my goal is to revive tugger as a project inside of the
PyOxidizer repository. Rather than import the code wholesale (it needs
a lot of work), my plan is to effectively cherry pick the good parts
of Tugger into the new code base. Eventually we'll get to a release
that we'll call version 0.2.

As part of this, we also add Sphinx scaffolding for Tugger.