starlark: move build path storage to build targets crate
This state is somewhat generic and shared among multiple logical
consumers. Let's move it to the new crate.

As part of this, we make resolution of the python distributions path
dynamic based on the current build path. Things probably should have
been this way all along, as there was the potential for things to get
out of sync.