tugger: move WiX code from PyOxidizer
Most of the code is commented out because it is missing types.

As part of this we also move some HTTP client code. This should have
been done in a separate commit. But I'm having a lazy Sunday.

There are tons of compiler warnings in tugger. Those will be
addressed in upcoming commits.