tugger: implement build() for WiXInstaller
This enables us to produce installers from Starlark!

As part of this, I encountered an error attempting to write the
WiX zip file into a directory that did not exist. As part of debugging
this, I introduced contexts to various errors so we can get a nice
chain of errors from build failures.