tugger: implement support for writing .deb files
This provides a rather low-level interface for writing .deb
files.

We haven't yet implemented support for compression. But it should
be easy enough.

Our deb writing code should be deterministic, which is a nice
property to have.

There are no tests and there are likely bugs. But at least the
old code from the old tugger crate mostly lives in the new code
base.