global: capture and display context when adding resources
There was a report on the Mercurial mailing list with a confusing error
message. We can improve the reporting of anyhow error messages by
formatting them with the debug formatter. This will emit any context
registered with the Error.

This commit changes a bunch of anyhow::Error conversions to use the
debug formatter. And we add contexts to some errors to improve
debugability.