pyoxidizer: add test building project with all config fields set
We have encountered a number of bugs related to the serializing of
the embedded Python interpreter config fields.

This commits attempts to put an end to the potential for bugs here by
adding a test that sets all fields and tests that we can build a project
with all fields serialized.

I wouldn't be surprised if there were still bugs here. But this
test coverage uncovered 2 bugs related to serializing strings. So
having the coverage is worth it.

This test inspired by the latest bug report in #309.