environment: remove runtime dependency on git2 crate
For reasons that are complicated, I'm blocked on importing the git2
crate to our monorepo at work, but I still want to be able to start
working on bazel rules for pyoxidizer binaries. This lets users avoid
the git2 dependency in the pyoxidizer tool so they can cut a
significant part of the dependency tree.