pyembed: use a new type to represent a resolved configuration
The configuration type currently self mutates as part of running. I'm
not keen on this pattern.

This commit introduces a new config type representing a resolved
configuration. When we pass a configuration into the interpreter, it
is converted to a resolved instance and that read-only data structure
is used as part of running the interpreter.

There is still some code that needs to be ported to the new type. But
this commit is a giant step in the right direction.