pyembed: always call Py_RunMain()
This regressed as part of removing the explicit `run_mode` field of
the Rust struct, which defaulted to running a repl. We had performed
this change in the 0.10 branch as commit
d2e351d57684addf7f0766955654932961612a81 but never forward ported the
fix to main.

Rather than preserve the previous API as we did with the aforementioned
commit, we rename the function to reflect that we always call
`Py_RunMain()` and finalize the interpreter.

After this change, we should start a REPL by default for the default
configuration file, like we did in 0.9 and 0.10.3.