pyembed: always run Py_RunMain()
This regressed as part of removing the explicit `run_mode` field of
the Rust struct, which defaulted to running a repl. We may want to
tweak the behavior in 0.11. But this is going to be an API incompatible
change. While I don't think people rely on semantic versioning for
pyembed, we might as well not get in the bad habit of breaking it.