pyoxidizer: add test demonstrating load() doesn't work
This effectively reproduced the failure reported in #328.

While I was here I added another test for register_target() working
using the full evaluation context because attempts to quickly fix
this resulted in that failing. So I want to be sure to get test
coverage in place before we change anything.