pyoxidizer: properly call build()
From the "I can't believe we shipped this" department...

Commit a84d6cf1fa3835fea2ab450bcb93739d50084086 was broken in multiple
ways. This commit aims to fix all of it.

The default implementation of has_attr() for Value will return an
Err if not explicitly implemented. The previous code resulted
in has_attr() resulting in error in scenarios where the test should
simply have been negative.

Furthermore, has_attr() isn't the proper method to call in the first
place, as the build method is part of the type object. So this commit
changes things to perform a lookup on the type value instead. That
at least gets us a working handle on the `Value` for the
method/function.

In addition, every call to build() would fail because we only passed
a single argument and omitted the `this` argument. So this commit
passes the resolved value to its `build()` method.

In addition, attempts to call build() would likely result in an error
due to nested borrowing of the context. This commit introduces scopes
to ensure the context is dropped before calling into another Starlark
function.

Altogether, we should now be able to build resolved targets again.
This exposes another gaping hole in test coverage. I sense a good
weekend project coming up...

Closes #329.