pyoxidizer: implement StarlarkEnvironment in terms of EvaluationContext
Previously, EvaluationContext and StarlarkEnvironment duplicated some
functionality for constructing and running a Starlark environment. The
code was out of sync and different in subtle ways. For example, we
didn't construct the context or config paths the same ways.

This commit changes StarlarkEnvironment to be implemented in terms
of EvaluationContext. This ensures we have a unified way of constructing
the Starlark environment. This will make future refactors simpler.