starlark: don't pass in file loader Environment
We always derive this from the tracked environment. So we can just
do that in the function instead of forcing the caller to clone the
Environment.

This enables us to get rid of the final consumer of the low-level
Starlark Environment! This should enable us to refactor how Environment
are used by our wrapper type.