pyoxidizer: use child environment for execution
Now that we've refactored how environments are initialized, we can
now freeze the parent environment and execute in a mutable child
environment.

This enables us to potentially execute multiple files or statements
in separate child environments. Although the API does not support this.

The big benefit of this change is that the `load()` Starlark function
now appears to work!