pyoxidizer: implement PythonExecutable.to_wix_bundle_builder()
This allows us to produce an exe installer which also installs the VC++
Redistributable. This provides a turnkey way for producing distributable
Windows installers for applications built with PyOxidizer.

TO get this working, I had to normalize some relative paths for some
reason to handle the special case of `.`. I'm unsure what is going on
here. We may want to perform more filename normalization in the virtual
file manifest code?