ci: use sccache
We install sccache using cargo and cache its binary to avoid this
overhead on subsequent invocations.

We start an sccache server before building and stop it afterwards,
printing stats before doing so.

We also set environment variables to point sccache at an S3 bucket
that I control. I manually installed secrets with an AWS access key
with read/write access to just the key prefix in this bucket.