ci: stop running `pyoxidizer --version`
I added this as part of debugging why the GitHub Actions environment
wasn't working. It turned out to be due to a shallow clone, which was
worked around several commits back.

Anyway, building the binary can add several dozen seconds to CI time.
IMO it isn't worth it. So let's stop incurring the overhead.