tugger: add function for locating Visual C++ Redistributable DLLs
Some applications may wish to ship vcruntimeXXX.dll next to their
binary instead of installing it from the official installers.

This commit adds a function for locating the DLLs so we can teach
packaging code to do this.

As part of this, we move an existing enum to the new windows module.