tugger: make Debian control file parsing more robust
This satisfied the inline TODO to parse the string value to determine
the field value type instead of relying on the field name. The new
approach is more resilient to unknown fields.

We found issues with the old approach by implementing a test that
parses package archive control files in the well-known directory for
such lists on Debian based distros. I am able to parse all the control
files on my machine with this change.