tugger: shore up control tar writing
We add support for writing extra files, which can be used to add
maintainer scripts. We also reusue the tar generation logic used for
data.tar files. The only reason not to would be if we had to emit the
`control` file first. But I don't believe this is the case.