tugger: finish moving analyze code out of pyoxidizer crate
This code is generic and should live in tugger. This code is quite
old and doesn't live up to my current Rust standards. But it can be
refactored in the future.

As part of this we refactored code to remove a dependency on itertools,
as I didn't want to introduce it into tugger since a lot of what you
can do with itertools you can do in the standard library these days.