tugger: add test for using signtool to sign an executable
We implemented the code a few commits back but weren't testing it.

In order to implement a test, we needed to pull in a crate to serialize
x509 certificates to pkcs #12 .pfx files. It might be possible to get
signtool to recognize other formats. I'm not sure. But .pfx seems to be
the most common. So it makes sense to support it.