tugger-debian: establish standalone crate
The tugger crate is growing quite large. Much of the code in the crate
is generic and can be used outside tugger itself. I anticipate people
will want to do this. My plan is to develop generic pieces of tugger
in their own domain-specific crates.

We start with the Debian code, as it is reasonably isolated and
free from other crate dependencies.