tugger-snapcraft: move snapcraft code to its own crate
Like we've done for other functionality. We want to move the
domain specific and generic code to its own crate to facilitate
reuse.

This required refactoring the Starlark code a bit since we can't
implement a trait on a type defined in another crate. In the
future, we might want to move the Starlark functionality to these
domain specific crates. But this will require a lot of refactoring.

As part of this, we moved glob functionality to tugger-common.