pyoxidizer: stop writing py-module-names
IIRC this was introduced to facilitate filtering modules and inspecting
results of packaging. We now have better mechanisms in place for both.
I don't think this will be missed much. If it is, there are better
ways to facilitate this, such as a command to analyze packed resources
in a file or binary.