pyoxidizer: misc changes to auto-generated main.rs file
We improve the documentation of how the config works. We also
annotate the type of the config variable so we're explicit. We
also document that py_runmain() finalizes the interpreter. We
also change the variable name for the exit code. We also tweak
the error message that is printed on interpreter construction
fail.