pyembed: transfer ownership of resources state into importer
Before, the importer state tracked whether ownership was transferred.
In the case of running an embedded Python interpreter, the importer
did not own the state and we stored resources state on the interpreter
instance.

After this commit, the importer is constructed by a
`Box<PythonResourcesState>` instead of a `&PythonResourcesState` and it
always owns the resources state instance.

We still have to use a capsule to store the raw pointer in the importer
state, as Python objects must have 'static lifetimes and the resources
lifetime cannot be 'static.