pyembed: add method to PythonResourcesState to load resources from a PyObject
Now that ownership of PythonResourcesState is transferred to the
importer, we can simplify how ownership of backing data works. In the
ideal world, PythonResourcesState would hold references to the data
structures backing its indexed data. So that's the direction we start
to move in.

This commit teaches PythonResourcesState to load packed resources data
contained in a PyObject. A reference to that PyObject is then stored
in PythonResourcesState to ensure it isn't garbage collected.

Code for instantiating an OxidizedFinder uses the new API. And the
importer state no longer holds a reference to a PyObject backing data.

A bonus side-effect of the new implementation is we can have multiple
PyObject backing indexed resources. Although this isn't yet tested.