pyembed: have test runner exit
I'm not sure when this regressed. But we were swallowing errors in
tests. Test failures were only visible from test output.

With the switch to use `py_runmain()` to invoke the embedded
interpreter and the switch to use rusty_fork_test!, I think it is safe
to have these tests exit on their own now.