pyoxidizer: define a load mode for packed resources
The new enum allows us to control whether packed resources are
embedded in the binary or serialized and loaded to a separate
file installed next to the binary.

We don't yet expose APIs to set this field. But the core functionality
now exists to build binaries that load resources from a standalone
file. And this will enable sharing resources file across binaries.