pyoxidizer: automatically index package license metadata
We just implemented code for scanning for license metadata. This
commit starts the process of wiring it up to packaging.

The plan for license scanning is to register all the license metadata
somewhere central. Then at some point during packaging (likely when
finalizing the resource collection), we'll validate that licenses are
in conformance with whatever policy we set.

This commit teaches the resource collector to store license metadata
for Python packages. The packaging primitives in PyOxidizer have been
updated to register license metadata with the resource collector
automatically.

Currently, the license metadata just sits there doing nothing.
Obviously this will change in future commits.

This change required us to make the various packaging APIs mutable,
as we now mutate the binary builder to store license metadata. This
had some fallout in Starlark code in order to appease the borrow
checker.