pyoxidizer: print report of license use when building
We're capturing license info on the resource collector. Let's start to
put it to use by exposing a report of license presence when building.

I'm sure things will change here. But I want to start doing useful
things with this metadata to flush out the direction we want to go in.

This partially addresses #142.