python-packaging: be more aggressive about finding top-level modules
It was suspicious that we failed to implement license annotation on
extensions and the license report said we were good. Upon further
inspection, it is important for us to annotate all top-level modules,
not just packages.

This commit changes our logic for identifying top-level elements so
we examine a larger set of resources for license annotations.
Extension modules now properly appear as not annotated.