pyembed: strip trailing .__init__ from module names when resolving resource
This gives us partial compatibility with CPython.

A more robust implementation would strip interior `__init__` from
module names as well as preserve a duplicate module with the
`__init__` naming variation. The former makes sense (if it occurs
in the wild - which I'm skeptical of, since `foo.__init__.bar` would
expand to a path like `foo/__init__/bar.py`, which shouldn't exist).
The latter is more nuanced. I would argue Python's behavior of
introducing a duplicate module object is wrong and an unintended
side-effect of the way its filesystem importer is implemented. So I
would prefer PyOxidizer to do something more reasonable here. But
at the same time, code may be relying on the current behavior and
we should arguably maintain compatibility with it.

Let's land this and see if anyone complains about the subtle
differences. Also, if the Python bug report ends up with the
current behavior being documented and part of the importer spec,
we should go along with their will at that point.

Closes #317.