pyembed: use Python's uuid module to generate modules filename
The `uuid` crate dependency was only for a likely lesser used feature.
I don't think the dependency is worth it. Let's use Python's uuid
module to generate the filename so we can remove the dependency.

A side effect of this is we no longer have any dependencies using the
BSD 3 clause license.