pyoxidizer: add a file containing licensing info for pyembed
We just implemented logic for generating a text file containing the
licensing information for the pyembed crate.

This commit vendors the output of that command and makes the content
available to Rust code, so the license text can be installed with a
built binary.

The release command has been updated to ensure the file matches
the current expected state.

The reason we don't just derive the license dynamically is because
it relies on functionality from `cargo-deny` and pulling in that
crate introduces a ton of dependencies. It's much easier to cache
the result in the repo.