tugger-licensing-net: extract network functions to own crate
I'm having a difficult time getting cargo deny properly picking
up the pruned crate dependencies for tugger-licensing. This is
preventing us from deriving the licensing information for the pyembed
crate properly.

As a workaround, this ocmmit introduces a separate crate to hold
the licensing functionality that requires a network connection.
I feel dirty for doing this. Maybe we can undo this someday...