pyembed: API: MainPythonInterpreter ownership prevents runtime errors
Previously, the `acquire_gil` method of a `MainPythonInterpreter` would return
an `Err` if called after `py_runmain`. Coordinating this required a state type
`InterpreterState`, which `MainPythonInterpreter` managed manually.

Now `py_runmain` takes ownership of its receiver, so Rust will refuse even to
compile a program that attempts to call `acquire_gil` after `py_runmain`.

Because such an error is now impossible, `acquire_gil` has become infallible.
I removed the `Result` wrapping of the `Python` return type to reflect this.

These are breaking changes, but `py_runmain`'s taking ownership of
`MainPythonInterpreter` is safer, taking advantage of the benefits of Rust that
make dealing with its borrow checker worthwhile. And if `acquire_gil` continued
to return a `Result`, it would be confusing for beginners, and the constant
`.unwrap()`ing would be un-ergonomic for maintainers.