pyoxidizer: upgrade to latest python-build-standalone distributions
This required various test changes due to changes in the distributions.

The license for the ncurses library is now properly annotated, resulting
in a Python extension no longer getting dropped in copyleft filtering
mode.

On macOS, we no longer link against libiconv, so extensions are no longer
dropped in copyleft filtering mode.