pyembed: move most memory allocator logic to pyalloc module
interpreter.rs is pretty gnarly. And the memory allocator code/logic
is currently spread across both it and pyalloc.rs.

This commit attempts to aggregate most of the allocator code into
pyalloc.rs. As part of this, we made the allocator enum/type a bit
more powerful so its internals are hidden from the caller.