pyembed: track allocator debug in a separate config field
The debug setting is independent of the memory domain used (it
applies to all of them). So it makes sense to track the allocator
debug bool as a separate config field.

We didn't support setting this before and don't yet support setting
it after this change. We'll implement that in subsequent commits.