pyembed: refactor allocator configuration
We now define a custom allocator backend in its own field and
have separate fields controlling how it is used. This ensures we
only have a single custom allocator defined rather than a mix of
multiple allocators.

We expose fields to enable installing custom allocators for the mem
and obj domains as well as pymalloc's arena allocator (although the
arena allocator is busted due to
https://github.com/dgrunwald/rust-cpython/issues/257).

We change the name of the "system" backend to "default" because that
is more appropriate.

I also found a subtle issue with the detection of the default backend
only checking for x86_64 on Windows: it should check for all Windows
MSVC arches.

Part of #358.