pyoxidizer: install mimalloc and snmalloc as Rust global allocator
Previously, use of jemalloc would make jemalloc the Rust global
allocator via the `jemallocator` crate. With the recent introduction
of support for mimalloc and snmalloc, we want to support a similar
feature.

This commit teaches the new project template to configure the Rust
global allocator for all 3 supported custom allocators. We do this
by introducing a new set of crate features for controlling the global
allocator explicitly. The autogenerated main.rs has explicit feature
checking and code for registering the global allocator. This is a
bit more verbose. But I like being explicit and consistent across the
allocators.

We've also introduced tests that verify autogenerated projects with
custom allocators compile correctly.

There is room to introduce Starlark config options to control the
global allocator settings. But these didn't exist before. So I think
we don't need to implement that functionality as part of this commit.

Both mimalloc and snmalloc use cmake as part of their crate build
process. So we document that and skip tests on Windows since cmake
isn't present in CI. We'll likely need more intelligent follow-ups
here (like to probe for the executable in the test). Let's wait for
things to break/complain before introducing more complexity.

Part of #358.