pyoxidizer: build with static CRT when using standalone_static distributions
I _think_ the old configuration of mixing the CRTs just happened to
work. Although it is unclear because the Windows static distributions
have been broken for so long. This commit teaches the build to
automatically use static CRT linkage when using the static
distributions. We also document the setting in the auto-generated
.cargo/config file so people know it may need set.