pyoxidizer: pass /FORCE:MULTIPLE when linking Windows static distributions
See the inline comments and
https://github.com/indygreg/python-build-standalone/issues/71 for more.

Even with the supposedly fixed python-build-standalone distributions,
ffi symbols are appearing in multiple object files and this leads to
a linker error. There are many symbols appearing multiple times
but the ffi symbols are the only ones the linker treats as an error.
Why, I'm unsure. The linker error message says the symbols are in
types.obj and cfield.obj. types.obj is coming from libffi itself
(likely via libffi.lib).

Anyway, I'm unsure what is going on here, but we can work around it by
telling the linker to be less strict. I'm not keen on this solution.
But it does enable the Windows static distributions to build. Therefore
we can close #360.