pyoxidizer: support copying VC Runtime DLLs next to built binary
tugger-windows and tugger-wix have functionality for locating the
VC++ Redistributable Runtime files. And PyOxidizer already had support
for plumbing this together and installing the runtime automatically
when converting a PythonExecutable to a WiX installer.

This commit builds on top of that work and teaches the generic
binary building code to install the runtime DLLs automatically, if
configured to do so.

Part of #353.