ci: build and test in release mode
We've added a lot of tests in recent months. It looks like we are
or soon will be at the point where the extra time spent building
in release mode if offset by the reduction in test run time.

TBH, I'm not sure how I feel about this. I feel like some of
the expensive test code could be optimized. But it is a non-release
build after all and I'm unsure whether it is worth doing that.
If we should be optimizing anything it should be release mode
builds.

There's potentially room to introduce or tweak the compiler
profiles to make debug builds run faster. Perhaps that would be
a better path forward?