tugger-apple: define useful SDK APIs
We add a helper function to convert the SDK version to a
semver::Version to facilitate easy sorting.

We also extract the logic for materializing an SDK from a directory
to its own function so it can be used in isolation. (Perhaps someone
already has the path to an SDK and wants to construct an AppleSdk
instance.)