tugger-apple-codesign: implement CLI for doing common things
I might as well create a CLI for this project since it will be
helpful for debugging and people will likely ask for it once the
crate implements useful functionality.

We start by implementing a command for extracting the CMS blob
from a Mach-O binary. We start here because that's the stage
I'm at with reimplementing code signing.