tugger-apple-codesign: return more info about Mach-O binary
This will allow us to reason about the existence of the __LINKEDIT
section with regards to the larger binary. This will be necessary
in order to write/mutate the __LINKEDIT section.

As part of this, we've defined a new sub-command for printing info
about the __LINKEDIT and signature data in a Mach-O binary.