tugger-apple-codesign: module for documenting format
The format and semantics of Apple code signing are not officially
documented and existing resources are often light on details. I
have to figure all this out as part of implementing code signing.
So I might as well document it so the next person trying to figure
this out knows how it works.