tugger-apple-codesign: start to implement signature builder
We want this crate to support generating signatures. This commit
starts the implementation of that. Things are far from complete but
I didn't want the commit to get too far away from me, as we're having
to start teaching blobs to serialize and hold their own storage.

I also made some changes to the captured linkedit metadata as part
of this, as I was debugging file layouts a bit.