tugger-apple-codesign: truncate blob slice when parsing SuperBlob
Before, extra bytes between blobs or after the last blob would be
assigned to a Blob in the index. Blob::from_bytes() implementations
would truncate the extra data. But this behavior was causing a
mismatch between the serialization of blobs and their original
attributed data because excess data was being attributes.

After this commit, the SuperBlob parser effectively strips excess
data from blob records. There is still room to "remember" where this
excess data was in case we need to preserve it. But that can be handled
at the SuperBlob / EmbeddedSignature layer.