cryptographic-message-syntax: API for signing data
This commit introduces an API for constructing an instance of
signed data. It allows you to progressively build up parameters
used for signing data.

A simple test demonstrating that RSA signatures can be produced
and verified has been added.

As part of this, we fixed a bug in SignedAttributes serializing
where it contained a redundant nested set.

We also fixed a bug where EncapsulatedContentInfo.eContent wasn't
using constructed values properly.