tugger-apple-codesign: start implementing signature verification
We want to verify that signatures and binaries we produce with this
crate are valid and usable. And, inspecting/auditing the state of
embedded signatures seems like a useful feature to have.

This commit introduces support for verifying binaries with embedded
signatures. Verification is far from comprehensive, but it is a start.
The new code has even revealed some bugs with our signing code! More
importantly, it has revealed some bugs in our signature parsing code
and/or the verification code. For example, code directories in the
wild have a special slot digest for resources. But we're not parsing
resources data, so the verifier thinks it is an extra digest.