tugger-apple-codesign: implement parsing of Code Resources XML plists
I'm not thrilled that we couldn't use Deserialize here. But I couldn't
figure out how to shoehorn Deserialize + Visitor implementation onto
the data stream that plist was emitting. I'm sure someone with more
knowledge of serde could figure it out.

I'm also not thrilled about the state of the data types here. The keys
in the XML format are kinda wonky and it isn't clear when something
goes in the original key of *2 of that key. I can make some guesses.
But more research/testing is needed.