tugger-apple-codesign: support setting code resources digest
When copying settings from an existing binary, we'll preserve the
existing digest, if present.

We also add an API to define the content of the code resources file
so it can be digested and added to the code directory.

And we teach the CLI to define a file to use. Although we don't
perform validation that the file is a well-formed XML plist.