tugger-apple-codesign: support emitting CMS signatures with CLI
As part of this I discovered that we weren't emitting the signature blob
properly. So that was fixed as part of this commit.

After this change, we're able to run `rcodesign
generate-self-signed-certificate` + `rcodesign sign --pem-source`
with the output of our self-signed certificate and a CMS signature is
materialized in the binary. ECDSA, ED25519, and RSA keys should all
work. But I've only tested the former 2, as ring doesn't have RSA key
generation.