docs: start to move PyOxidizer docs to pyoxidizer/docs
We have multiple Sphinx projects in the repo. We want each project to
have its own docs and the docs should live under that project.

Currently, PyOxidizer's docs mostly live in docs/. This commit
marks the start of moving those docs to pyoxidizer/docs.

We start by registering `pyoxidizer` as a project prefix. We then
move existing files with this prefix to pyoxidizer/docs. There should
be no change in behavior as a result of this commit.