tugger-apple-codesign: start to unify error type
We have a lot of errors in this crate. I want to unify them. This
commit starts the process.

We begin with errors from the CLI, since it is the outer-most error
type.

We also employ the use of thiserror, as it is a popular crate and
helps us avoid error type boilerplate.