tugger-apple-codesign: sign main executable in a bundle
Previously we handled all files except the main executable, which is
special because it contains a hash of the resources file and Info.plist.

This commit teaches the bundle signer to handle the main executable.

I haven't tested this thoroughly and there's still a lot of work to
be done in bundle signing land to get all the desired signing options
to apply. But after this commit, the main executable is signed and
installed and a signed MacOS App Bundle should contain all files there
were present in the source bundle.