tugger-apple-codesign: always remove Contents/ prefix from resources path
We need to remove the prefix for the stored path in addition to the
rule-processed path to match behavior of `codesign`.

With this change, our CodeResources files are now bit identical to
`codesign`s except in cases where we rewrite Mach-O binaries!
`codesign --verify` says sealed resources are happy in most cases.
But there are still verification issues due to nested bundles and
Mach-O binaries.