tugger-apple-codesign: use truncated SHA-256 for CodeResources cdhash
This appears to be what `codesign` does by default. Although it appears
that SHA-1 works just as well. But SHA-1 isn't as secure.

I'll have to dig into what versions of operating systems support
SHA-256. For now, it seems like a reasonable default. But we may
have to make it configurable for signature compatibility. Time will
tell.