tugger-apple-codesign: use SigningSettings to control signing behavior
This commit does a lot and its scope got away from me.

We refactor how code signing settings work so there is a single
type holding settings. That type is passed around to the various
signers who look up settings from it at signing time.

Signing settings are "scoped." We can have global settings or granular
settings that only apply to certain paths or even certain Mach-O
instances within a fat binary. As signers traverse into nested elements,
they coerce the settings into a new instance that applies to the
entity they are operating on. This is a somewhat complex solution and
there are likely bugs. But the complexity seems warranted since
people will need to target specific settings to specific entities in
the wild.