release: further refine commit relevancy detection
We stardardize on using `releasebot: <action>` as the prefix for all
commit messages so they are machine readable.

We refine the logic for filtering commits to determine if they are
relevant and necessitate a version bump.

In no-release mode this appears to do the right thing. Although it is
really difficult to tell without actually doing a release.