pyembed: rewrite the path hook filtering logic
One thing I've learned from hacking on version control tools is that
path canonicalization is really hard and if I had a time machine I
would be the use of things like symlinks in certain contexts because
getting things correct is really hard.

After thinking about how path hooks are implemented, I feel the old way
of accepting various types and canonicalizing paths is too complex.
The string encoding is a nightmare and it is much easier to not have
to worry about it. Furthermore, canonicalizing at run-time is kind of
brittle, as it relies on resolving symlinks and incurring I/O. This
adds overhead and behavior can change as the program is executing due
to the filesystem changing.

This commit simplifies the implementation by only accepting str path
values. We also limit the set of accepted values to be either identical
to current_exe (not sys.executable) or having a strict prefix thereof.
We also replace the path based logic with string based logic, taking
care to reject and test for edge cases involving directory separators
and relative paths.