pyembed: change how name filtering is done for iter_modules()
We recently introduced is_visible() to both the generic finder and the
path specific finder cases. I want to change the behavior of
is_visible() but couldn't easily do this since the iter_modules()
filtering was so tightly coupled to it.

This commit changes the filtering callback so it not only filters but
also does name translation.

As part of this, OxidizedFinder no longer calls into is_visible()
and instead performs its own trivial filtering based on presence of
dots.

Test behavior doesn't change, which is a good sign.