pyoxidizer: capture pyembed crate version in environment.rs
Previously, we assumed that the version of the pyembed crate exactly
matched PyOxidizer's. This was a faulty assumption, as our release
automation now supports skipping crate releases if there are no
meaningful changes.

This commit introduces a constant defining the pyembed crate's
version string and uses that constant when materializing a Cargo.toml
fragment for the pyembed crate.

The release automation has been taught to update environment.rs
whenever the crate's version changes.

This should hopefully prevent #382 from happening again.