tugger-rust-toolchain: implement functionality for downloading Rust toolchains
This is like a mini rustup implementation. Not the fastest or the best
code I've ever written. But it gets the job done. It appears to parse
manifests and download package archives just fine.

Impetus for this work is so PyOxidizer can automatically install
Rust so end-users who aren't Rust developers don't have to worry
about doing this themselves.