pyoxidizer: start to use managed Rust installation
PyOxidizer's target audience is Python developers, not Rust developers.
Rust is an implementation detail of PyOxidizer and that implementation
detail should ideally be hidden from most users.

This commit teaches PyOxidizer to manage the Rust toolchain
automatically. Using the tugger-rust-toolchain (which provides
rustup-like functionality), we hardcode the Rust toolchain we want
to use and teach PyOxidizer to obtain and use it automatically.

There's still a bit of tweaking we need to do here, including
optimization (tests in debug builds are super slow), documentation,
and support for overriding the default. We'll address these in future
commits.