pyembed: update test to account for cpython 0.6 changes
cpython 0.6 now internally stores strings as different types
if they aren't UTF-8. This introduced a test failure on Windows.

As part of debugging this, I realized that we defined our UTF-16
string on Windows in the wrong endian. So the tests were changed
to fix this.

Tests were also simplified, as we can now access the internal
data stored in the Python object without having to go through
unsafe APIs.