pyembed: initialize test interpreters with argv set to actual interpreter
This forces Python path detection to work as `python` would.

The effect of this is that Python is able to pick up its own resource
paths in more scenarios than before, leading to fewer test failures
due to interpreter initialization failure.

Because we're now setting argv and needs argv[0] to be a special
value, we had to adjust various tests accordingly.